refac: litellm
